# openapi.yaml
openapi: 3.0.3
info:
  title: Sax Buddy Service API
  description: |
    REST backend service for a multi-agent AI application for saxophone learning.
    This system processes structured performance data from mobile devices to provide 
    adaptive feedback and generate personalized lesson plans based on multi-dimensional 
    skill tracking (intonation, rhythm, articulation, dynamics).
  version: 0.1.0
  contact:
    name: Sax Buddy Team
  license:
    name: MIT
servers:
  - url: https://api.saxbuddy.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
paths:
...
  /users:
    post:
      tags: [Users]
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
  /users/{user_id}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{user_id}/profile:
    get:
      tags: [Users]
      summary: Get user profile
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    
    put:
      tags: [Users]
      summary: Update user profile
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /users/{user_id}/progress:
    get:
      tags: [Users]
      summary: Get user progress
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User progress data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProgress'

  # Performance Tracking
  /performance/sessions:
    post:
      tags: [Performance]
      summary: Start new practice session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerformanceSessionCreate'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceSession'
  /performance/sessions/{session_id}:
    get:
      tags: [Performance]
      summary: Get session details
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceSession'
    
    patch:
      tags: [Performance]
      summary: Update session (e.g., end session)
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerformanceSessionUpdate'
      responses:
        '200':
          description: Session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceSession'
...
  /exercises:
    get:
      tags: [Content]
      summary: List exercises
      parameters:
        - name: exercise_type
          in: query
          schema:
            $ref: '#/components/schemas/ExerciseType'
        - name: difficulty_level
          in: query
          schema:
            $ref: '#/components/schemas/DifficultyLevel'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exercise'
  /exercises/{exercise_id}:
    get:
      tags: [Content]
      summary: Get exercise details
      parameters:
        - $ref: '#/components/parameters/ExerciseId'
      responses:
        '200':
          description: Exercise details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
  /users/{user_id}/lesson-plans:
    get:
      tags: [Content]
      summary: Get user's lesson plans
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: is_active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: User's lesson plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LessonPlan'

    post:
      tags: [Content]
      summary: Generate new lesson plan for user
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '201':
          description: Lesson plan generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonPlan'
  /lesson-plans/{plan_id}/lessons:
    get:
      tags: [Content]
      summary: Get lessons in plan
      parameters:
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200':
          description: Lessons in plan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
  /lessons/{lesson_id}:
    get:
      tags: [Content]
      summary: Get lesson details
      parameters:
        - $ref: '#/components/parameters/LessonId'
      responses:
        '200':
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
    
    patch:
      tags: [Content]
      summary: Update lesson completion status
      parameters:
        - $ref: '#/components/parameters/LessonId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonUpdate'
      responses:
        '200':
          description: Lesson updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'

  # Assessment & Feedback
  /users/{user_id}/assessments:
    get:
      tags: [Assessment]
      summary: Get user's formal assessments
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: User's assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FormalAssessment'

    post:
      tags: [Assessment]
      summary: Trigger formal assessment
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentTrigger'
      responses:
        '201':
          description: Assessment completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormalAssessment'
  /performance/sessions/{session_id}/feedback:
    get:
      tags: [Assessment]
      summary: Get session feedback
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
...
components:
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ExerciseId:
      name: exercise_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PlanId:
      name: plan_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    LessonId:
      name: lesson_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Resource not found"
  schemas:
...
    SkillLevel:
      type: string
      enum: [beginner, intermediate, advanced]
    PracticeFrequency:
      type: string
      enum: [daily, weekly, occasional]
    ExerciseType:
      type: string
      enum: [scales, arpeggios, technical, etudes, songs, long_tone]
    DifficultyLevel:
      type: string
      enum: [beginner, intermediate, advanced]
    SessionStatus:
      type: string
      enum: [in_progress, completed, paused, cancelled]
    AssessmentType:
      type: string
      enum: [periodic, on_demand, milestone]
    TriggerReason:
      type: string
      enum: [scheduled_interval, user_requested, skill_threshold, lesson_completion]
    FeedbackType:
      type: string
      enum: [post_session, motivational, corrective]

    # User Domain
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, email, name, is_active, created_at, updated_at]
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
      required: [email, name]
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        current_skill_level:
          $ref: '#/components/schemas/SkillLevel'
        learning_goals:
          type: array
          items:
            type: string
        practice_frequency:
          $ref: '#/components/schemas/PracticeFrequency'
        formal_assessment_interval_days:
          type: integer
          minimum: 1
        preferred_practice_duration_minutes:
          type: integer
          minimum: 1
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        current_skill_level:
          $ref: '#/components/schemas/SkillLevel'
        learning_goals:
          type: array
          items:
            type: string
        practice_frequency:
          $ref: '#/components/schemas/PracticeFrequency'
        formal_assessment_interval_days:
          type: integer
          minimum: 1
        preferred_practice_duration_minutes:
          type: integer
          minimum: 1
          nullable: true
    UserProgress:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        intonation_score:
          type: number
          minimum: 0
          maximum: 100
        rhythm_score:
          type: number
          minimum: 0
          maximum: 100
        articulation_score:
          type: number
          minimum: 0
          maximum: 100
        dynamics_score:
          type: number
          minimum: 0
          maximum: 100
        overall_score:
          type: number
          minimum: 0
          maximum: 100
        sessions_count:
          type: integer
          minimum: 0
        total_practice_minutes:
          type: integer
          minimum: 0
        last_formal_assessment_date:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # Performance Domain
    PerformanceSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        exercise_id:
          type: string
          format: uuid
        duration_minutes:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/SessionStatus'
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PerformanceSessionCreate:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        exercise_id:
          type: string
          format: uuid
      required: [user_id, exercise_id]
    PerformanceSessionUpdate:
      type: object
      properties:
        duration_minutes:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/SessionStatus'
        ended_at:
          type: string
          format: date-time
        notes:
          type: string
...
    Exercise:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        exercise_type:
          $ref: '#/components/schemas/ExerciseType'
        difficulty_level:
          $ref: '#/components/schemas/DifficultyLevel'
        estimated_duration_minutes:
          type: integer
          minimum: 1
        instructions:
          type: array
          items:
            type: string
        reference_audio_url:
          type: string
          format: uri
          nullable: true
        sheet_music_url:
          type: string
          format: uri
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LessonPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        target_skill_level:
          $ref: '#/components/schemas/DifficultyLevel'
        estimated_completion_days:
          type: integer
          minimum: 1
        lesson_ids:
          type: array
          items:
            type: string
            format: uuid
        is_active:
          type: boolean
        generated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        next_formal_assessment_due:
          type: string
          format: date-time
          nullable: true
    Lesson:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lesson_plan_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        order_in_plan:
          type: integer
          minimum: 1
        exercise_ids:
          type: array
          items:
            type: string
            format: uuid
        estimated_duration_minutes:
          type: integer
          minimum: 1
        learning_objectives:
          type: array
          items:
            type: string
        prerequisites:
          type: array
          items:
            type: string
        is_completed:
          type: boolean
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    LessonUpdate:
      type: object
      properties:
        is_completed:
          type: boolean
        completed_at:
          type: string
          format: date-time

    # Assessment Domain
    FormalAssessment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        assessment_type:
          $ref: '#/components/schemas/AssessmentType'
        trigger_reason:
          $ref: '#/components/schemas/TriggerReason'
        skill_metrics:
          type: object
          additionalProperties:
            type: number
        overall_score:
          type: number
          minimum: 0
          maximum: 100
        skill_level_recommendation:
          $ref: '#/components/schemas/SkillLevel'
        improvement_areas:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        next_lesson_plan_id:
          type: string
          format: uuid
          nullable: true
        assessed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    AssessmentTrigger:
      type: object
      properties:
        trigger_reason:
          $ref: '#/components/schemas/TriggerReason'
      required: [trigger_reason]
    Feedback:
      type: object
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        feedback_type:
          $ref: '#/components/schemas/FeedbackType'
        message:
          type: string
        encouragement:
          type: string
        specific_suggestions:
          type: array
          items:
            type: string
        areas_of_improvement:
          type: array
          items:
            type: string
        strengths_highlighted:
          type: array
          items:
            type: string
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        created_at:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...